Add support for 'Delivery' entities for folder creation
Currently it is not possible to add a delivery folder to the schema as there is no 'code' field for a delivery entity. 
Either the default fields for a Delivery needs to be updated to include a 'code' rather than 'title' field, or we need to add support to entity.py to provide a suitable field to use for naming delivery folders; eg the 'title' field.